+++ src/NeuralNet/Initializers/He.php
@@ -33,8 +33,7 @@ class He implements Initializer
      */
     public function initialize(int $fanIn, int $fanOut) : Matrix
     {
-        $scale = (6. / ($fanIn + $fanOut)) ** self::ETA;
-
-        return Matrix::uniform($fanOut, $fanIn)->multiply($scale);
+        return Matrix::uniform($fanOut, $fanIn)
+            ->multiply((6. / ($fanOut + $fanIn)) ** self::ETA);
     }
 }