+++ src/NeuralNet/Layers/Dense.php
@@ -197,8 +197,8 @@ class Dense implements Hidden, Parametric
 
         $w = $this->weights->w();
 
-        $optimizer->step($this->weights, $dW);
-        $optimizer->step($this->biases, $dB);
+        $this->weights->update($optimizer->step($this->weights, $dW));
+        $this->biases->update($optimizer->step($this->biases, $dB));
 
         unset($this->input);