+++ src/NeuralNet/Optimizers/AdaGrad.php
@@ -53,7 +53,7 @@ class AdaGrad implements Optimizer, Adaptive
         $this->rate = $rate;
     }
     /**
-     * Warm the cache with a parameter.
+     * Warm the cache.
      *
      * @param \Rubix\ML\NeuralNet\Parameters\Parameter $param
      */
@@ -67,8 +67,9 @@ class AdaGrad implements Optimizer, Adaptive
      *
      * @param \Rubix\ML\NeuralNet\Parameters\Parameter $param
      * @param \Rubix\Tensor\Tensor $gradient
+     * @return \Rubix\Tensor\Tensor
      */
-    public function step(Parameter $param, Tensor $gradient) : void
+    public function step(Parameter $param, Tensor $gradient) : Tensor
     {
         $norm = $this->cache[$param->id()];
 
@@ -76,9 +77,7 @@ class AdaGrad implements Optimizer, Adaptive
 
         $this->cache[$param->id()] = $norm;
 
-        $step = $gradient->multiply($this->rate)
+        return $gradient->multiply($this->rate)
             ->divide($norm->sqrt()->clipLower(EPSILON));
-
-        $param->update($step);
     }
 }