+++ src/NeuralNet/Optimizers/Stochastic.php
@@ -43,9 +43,10 @@ class Stochastic implements Optimizer
      *
      * @param \Rubix\ML\NeuralNet\Parameters\Parameter $param
      * @param \Rubix\Tensor\Tensor $gradient
+     * @return \Rubix\Tensor\Tensor
      */
-    public function step(Parameter $param, Tensor $gradient) : void
+    public function step(Parameter $param, Tensor $gradient) : Tensor
     {
-        $param->update($gradient->multiply($this->rate));
+        return $gradient->multiply($this->rate);
     }
 }