+++ tests/NeuralNet/Optimizers/RMSPropTest.php
@@ -2,12 +2,15 @@
 
 namespace Rubix\ML\Tests\NeuralNet\Optimizers;
 
+use Rubix\Tensor\Tensor;
 use Rubix\Tensor\Matrix;
 use Rubix\ML\NeuralNet\Optimizers\RMSProp;
 use Rubix\ML\NeuralNet\Optimizers\Adaptive;
 use Rubix\ML\NeuralNet\Optimizers\Optimizer;
+use Rubix\ML\NeuralNet\Parameters\Parameter;
 use Rubix\ML\NeuralNet\Parameters\MatrixParam;
 use PHPUnit\Framework\TestCase;
+use Generator;
 
 class RMSPropTest extends TestCase
 {
@@ -25,30 +28,36 @@ class RMSPropTest extends TestCase
         $this->assertInstanceOf(Optimizer::class, $this->optimizer);
     }
 
-    public function test_initialize_step()
+    /**
+     * @dataProvider step_provider
+     */
+    public function test_warm_step(Parameter $param, Tensor $gradient, array $expected)
     {
-        $param = new MatrixParam(Matrix::quick([
-            [0.1, 0.6, -0.4],
-            [0.5, 0.6, -0.4],
-            [0.1, 0.1, -0.7],
-        ]));
-
-        $gradient = Matrix::quick([
-            [0.01, 0.05, -0.02],
-            [-0.01, 0.02, 0.03],
-            [0.04, -0.01, -0.5],
-        ]);
-
-        $expected = [
-            [0.09683772233983162, 0.5968377223398316, -0.39683772233983167],
-            [0.5031622776601684, 0.5968377223398316, -0.4031622776601684],
-            [0.09683772233983162, 0.10316227766016839, -0.6968377223398315],
-        ];
-
         $this->optimizer->warm($param);
-        
-        $this->optimizer->step($param, $gradient);
 
-        $this->assertEquals($expected, $param->w()->asArray());
+        $step = $this->optimizer->step($param, $gradient);
+
+        $this->assertEquals($expected, $step->asArray());
+    }
+
+    public function step_provider() : Generator
+    {
+        yield [
+            new MatrixParam(Matrix::quick([
+                [0.1, 0.6, -0.4],
+                [0.5, 0.6, -0.4],
+                [0.1, 0.1, -0.7],
+            ])),
+            Matrix::quick([
+                [0.01, 0.05, -0.02],
+                [-0.01, 0.02, 0.03],
+                [0.04, -0.01, -0.5],
+            ]),
+            [
+                [0.0031622776601683794, 0.003162277660168379, -0.0031622776601683794],
+                [-0.0031622776601683794, 0.0031622776601683794, 0.0031622776601683794],
+                [0.0031622776601683794, -0.0031622776601683794, -0.0031622776601683794],
+            ],
+        ];
     }
 }