+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     - Renamed Popularity Contest strategy to Prior
     - Datasets now inherit from abstract parent Dataset class
     - Removed Dataset interface
+    - Neural net parameter update in Layer instead of Optimizer
 
 - 0.0.14-beta
     - Added feature importances to Gradient Boost